Moved mipro_optimizer from default imports
